dtor for Manager now destroys it's engines.
As a small ID10T guard against fouling up later versions of
CreateEngine() and not catching it, ~Manager asserts that the delete
functions are not null.

The whole point of having Development/Release/Debug configurations, is
so that there is a convenient middle ground for testing without the
overhead of debugging symbols, nor including !Release code in Release
builds.