Added a handle like class for native types, Cassius::Value.
Since the interface to calling script functions has been
defined in terms of a stack of C/C++ style values, it is much more direct to
be able to use a stack of (potentionally) disparate types for mapping
C++ argument lists to something that script code can call. I.e. for
exporting functions to scripts.

That's what this class is for: making it possible to move values between
script code and C++ code.  While arguably not the worlds best way of
doing it, this does get the job done.

An alternative would have been to use one of tr1::tuple /
boost::tuples::tuple / std::tutple, respectively of TR1, Boost, and
C++0x to move more work down into compile time. However Boost is a big
ass dependency, and support for TR1/C++0X support cannot be *required*
by Cassius. Maybe in another ten years or so...