Bootstrap CMake support.
This loosely resembles the old premake4 build.

cxx-plugin is rolled up to the latest and treated as an object library, not
that it makes much difference.

A copy of the Engine.tpp that premake4 generates is added to compile.

Some support code for guessing a vcpkg like triplet, and being able
to apply the toolchain file for VCPKG_ROOT is added. Easier to go that
route for the Windows build.

Still plenty to do, but enough to get compilation up for the further
development on Cassius.