Add Engine.tpp.
Committing for ease of compiling. But this may simply get removed
later on. It's only here because a decade ago the compilers I was
supporting couldn't handle doing this in sane way(tm), and I chose
to toss Lua at the problem.

As the .gitignore used to say,  to support crappy compilers.