Use cmake's GenerateExportHeader module.
Cassius.hpp now defines CASSIUS_API as an alias for CASSIUS_EXPORT.

The language plugins now require Cassius_EXPORTS to be defined
when they are compiled, so that the correct dllimport / dllexport stuff
is used. CMake automagically adds {targetname}_EXPORTS but keeps
these private, so e.g. CassiusCluaEngine, etc, do not pick up the
compile definition. An alternative might be to use a public
target_compile_definitions and a BUILD_INTERFACE generator
expression on the main Cassius target.

Using the project's binary dir as the place to generate an
Include/Cassius to store the export header. So that it's fairly
uniform across source modules. Kindly, cmake seems to pick this
up as a public include as part of the BUILD_INTERFACE generator ;).