Kludge for the joys of Windows.
On Linux cmake generally makes use of the linker's rpath to make
these kinds of things Not So Painful.

On Windows it looks like we need to copy the Lua DLL next to
the executable or the plugin. Otherwise the test program will
load the plugin using %PATH% and then LoadLibrary() will fail with
error code 126/module not found because it's not finding the Lua DLL
from %PATH% like it found the plugin.

On the flipside this makes running the test suite much less annoying.