Rewrite Manager.
This changes from the ugly hard coded backends structure to using
a set of Cassius::Plugin. Not exactly fond of the one Engine to Plugin
bit but it should be a minor problem for unix like and NT based targets.

To keep track of cleanup a mapping of the returned Engine pointers to
their owning Plugins are kept. Along with there finally being a method
to destroy the Engine pointers without destroying the Manager.

I'd like to see this done a bit better rather than relying on the ref
counting of dlopen() / LoadLibrary(), but it should do for now.