Support wrapping a script's function for C++ use.
Ahh, the beauty of time. Namely ten years later the compilers that
made it so troublesome to do this are now so old that"IDGA*" and
the feature is wrapped in a conditional compilation for modern C++.

Wrapping as C++14, as I find i thard to remember which bits and
pieces of syntax land in 11 and 14, and by the time the compiler isn't
a problem you're basically at C++14.