Remove the old backends enum.
When Cassius was originally written, v8 was young, hot, sexy; and
Spidermonkey while a pain in the rear to compile was readily available
for BSD. So I went with flexibility and started working on a Spidermonkey
backend. Sadly, it got lost to the test of time and unknown which
backups it landed in.

At this point? I don't think it is necessary to control which backend is
loaded for a target language. In the sense that the focus should be the
language not the backend, and a better approach would be leverage the
plugin system so that the choice could be overriden by user written
plugins.

The possibility of hooking in a user written plugin is probably the only
good reason not to move Cassius away form the old plugin loading
thing and into a regular shared library format.